+++ back/prisma/migrations/84_add_bsff_transporter_plates.sql
@@ -1,0 +1,2 @@
+ALTER TABLE "default$default"."Bsff"
+  ADD COLUMN "transporterTransportPlates" TEXT[];